// grayjay-android/app/src/unstable/assets/sources/bilibili/src/types.ts
export type RequestMetadata<X> = {
    request(builder: BatchBuilder): BatchBuilder
    process(http_response: BridgeHttpResponse<string>): X
}

export type Params = {
    readonly [key: string]: string
}
// ...
export type LoggedInNavResponse = {
    readonly data: {
        readonly isLogin: true
        readonly mid: number
        readonly face: string
        readonly uname: string
    }
}

export type UserSubscriptionsResponse = {
    readonly data: {
        readonly list: {
            readonly mid: number
        }[]
        readonly total: number
    }
}
// ...
export type WatchLaterResponse = {
    readonly data: {
        readonly count: number
        readonly list: {
            readonly pic: string
            readonly title: string
            readonly pubdate: number
            readonly owner: {
                readonly mid: number
                readonly name: string
                readonly face: string
            }
            readonly duration: number
            readonly bvid: string
            readonly cid: number
            readonly stat: {
                readonly view: number
                readonly like: number
            }
        }[]
    }
}
// ...
export type SpaceFavoritesResponse = {
    /** Null if there is no favorites section */
    readonly data: null | {
        /** Null if the count of favorites lists is 0 */
        list: null | {
            readonly media_count: number
            readonly title: string
            readonly id: number
        }[]
        readonly count: number
    }
}